split unet.py file
- move NN layer functions to layers.py
  - split conv_pass into conv_pass and conv
  - conv: single convolution + takes care of dimensions
  - conv_pass: block of convoltions, optional batch norm and
    residual connection
  - upsampling: add upsampling arg
    - replaces constant_upsample to allow more options
    - type of upsampling used, one of
      ['transposed_conv', 'resize_conv', 'uniform_transposed_conv']
    - transposed_conv: use a normal transposed convolution
    - resize_conv: replicate each pixel `factors` times followed
      by a 1x1 conv
    - uniform_transposed_conv: use a transposed convolution with
      a factors sized filter with a single, learnable value at all
      positions
- move cropping functions to utils.py
  - rename crop to crop_spatial_temporal
  - move (all-dimension) crop function from mknet to utils.py
