+++ funlib/learn/tensorflow/models/unet.py
@@ -1,499 +1,13 @@
-from .conv4d import conv4d
-from .utils import get_number_of_tf_variables
-import math
+import logging
+
 import numpy as np
 import tensorflow as tf
 
-
-def conv_pass(
-        fmaps_in,
-        kernel_sizes,
-        num_fmaps,
-        activation='relu',
-        name='conv_pass',
-        fov=(1, 1, 1),
-        voxel_size=(1, 1, 1),
-        padding='valid'):
-    '''Create a convolution pass::
-
-        f_in --> f_1 --> ... --> f_n
-
-    where each ``-->`` is a convolution followed by a (non-linear) activation
-    function. One convolution will be performed for each entry in
-    ``kernel_sizes``. Each convolution will decrease the size of the feature
-    maps by ``kernel_size-1``.
-
-    Args:
-
-        f_in:
-
-            The input tensor of shape ``(batch_size, channels, [length,] depth,
-            height, width)``.
-
-        kernel_sizes:
-
-            Sizes of the kernels to use. Forwarded to tf.layers.conv3d.
-
-        num_fmaps:
-
-            The number of feature maps to produce with each convolution.
-
-        activation:
-
-            Which activation to use after a convolution. Accepts the name of
-            any tensorflow activation function (e.g., ``relu`` for
-            ``tf.nn.relu``).
-
-        name:
-
-            Base name for the conv layer.
-
-        fov:
-
-            Field of view of fmaps_in, in physical units.
-
-        voxel_size:
-
-            Size of a voxel in the input data, in physical units.
-
-        padding:
-
-            'valid' or 'same', controls the padding on the convolution
-
-    Returns:
-
-        (fmaps, fov):
-
-            The feature maps after the last convolution, and a tuple
-            representing the field of view.
-    '''
-
-    fmaps = fmaps_in
-    if activation is not None:
-        activation = getattr(tf.nn, activation)
-
-    for i, kernel_size in enumerate(kernel_sizes):
-        in_shape = tuple(fmaps.get_shape().as_list())
-
-        # Explicitly handle number of dimensions
-        if len(in_shape) == 6:
-            conv_op = conv4d
-        elif len(in_shape) == 5:
-            conv_op = tf.layers.conv3d
-        elif len(in_shape) == 4:
-            conv_op = tf.layers.conv2d
-        else:
-            raise RuntimeError(
-                "Input tensor of shape %s not supported" % (in_shape,))
-
-        if isinstance(kernel_size, int):
-            kernel_size = [kernel_size]*(len(in_shape) - 2)
-
-        fov = tuple(
-            f + (k - 1)*vs
-            for f, k, vs
-            in zip(fov, kernel_size, voxel_size)
-        )
-
-        fmaps = conv_op(
-            inputs=fmaps,
-            filters=num_fmaps,
-            kernel_size=kernel_size,
-            padding=padding,
-            data_format='channels_first',
-            activation=activation,
-            name=name + '_%i' % i)
-
-        out_shape = tuple(fmaps.get_shape().as_list())
-
-        # eliminate t dimension if length is 1
-        if len(out_shape) == 6:
-            length = out_shape[2]
-            if length == 1:
-                out_shape = out_shape[0:2] + out_shape[3:]
-                fmaps = tf.reshape(fmaps, out_shape)
-
-    return fmaps, fov
-
-
-def downsample(
-        fmaps_in,
-        factors,
-        name='down',
-        voxel_size=(1, 1, 1)):
-    voxel_size = tuple(vs*fac for vs, fac in zip(voxel_size, factors))
-    in_shape = fmaps_in.get_shape().as_list()
-
-    # Explicitly handle number of dimensions
-    is_4d = len(in_shape) == 6
-    is_2d = len(in_shape) == 4
-
-    if is_4d:
-        orig_in_shape = in_shape
-        # store time dimension in channels
-        fmaps_in = tf.reshape(fmaps_in, (
-            in_shape[0],
-            in_shape[1]*in_shape[2],
-            in_shape[3],
-            in_shape[4],
-            in_shape[5]))
-        in_shape = fmaps_in.get_shape().as_list()
-
-    if not np.all(np.array(in_shape[2:]) % np.array(factors) == 0):
-        raise RuntimeWarning(
-            "Input shape %s is not evenly divisible by downsample factor %s." %
-            (in_shape[2:], factors))
-
-    if is_2d:
-        fmaps = tf.layers.max_pooling2d(
-            fmaps_in,
-            pool_size=factors,
-            strides=factors,
-            padding='valid',
-            data_format='channels_first',
-            name=name,
-        )
-    else:
-        fmaps = tf.layers.max_pooling3d(
-            fmaps_in,
-            pool_size=factors,
-            strides=factors,
-            padding='valid',
-            data_format='channels_first',
-            name=name)
-
-    if is_4d:
-
-        out_shape = fmaps.get_shape().as_list()
-
-        # restore time dimension
-        fmaps = tf.reshape(fmaps, (
-            orig_in_shape[0],
-            orig_in_shape[1],
-            orig_in_shape[2],
-            out_shape[2],
-            out_shape[3],
-            out_shape[4]))
-
-    return fmaps, voxel_size
-
-
-def repeat(
-        fmaps_in,
-        multiples):
-
-    expanded = tf.expand_dims(fmaps_in, -1)
-    tiled = tf.tile(expanded, multiples=(1,) + multiples)
-    repeated = tf.reshape(tiled, tf.shape(fmaps_in) * multiples)
-
-    return repeated
-
-
-def upsample(
-        fmaps_in,
-        factors,
-        num_fmaps,
-        activation='relu',
-        name='up',
-        voxel_size=(1, 1, 1),
-        constant_upsample=False):
-    '''Upsample feature maps with the given factors using a transposed
-    convolution.
-
-    Args:
-
-        fmaps_in (tensor):
-
-            The input feature maps of shape `(b, c, d, h, w)`. `c` is the
-            number of channels (number of feature maps).
-
-        factors (``tuple`` of ``int``):
-
-            The spatial upsampling factors as `(f_z, f_y, f_x)`.
-
-        num_fmaps (``int``):
-
-            The number of output feature maps.
-
-        activation (``string``):
-
-            Which activation function to use.
-
-        name (``string``):
-
-            Name of the operator.
-
-        voxel_size (``tuple`` of ``int``, optional):
-
-            Voxel size of the input feature maps. Used to compute the voxel
-            size of the output.
-
-        constant_upsample (``bool``, optional):
-
-            Whether to restrict the transpose convolution kernels to be
-            constant values. This might help to reduce checker board artifacts.
-
-    Returns:
-
-        `(fmaps, voxel_size)`, with `fmaps` of shape `(b, num_fmaps, d*f_z,
-        h*f_y, w*f_x)`.
-    '''
-
-    # Explicitly handle number of dimensions
-    is_2d = len(fmaps_in.get_shape().as_list()) == 4
-
-    voxel_size = tuple(vs/fac for vs, fac in zip(voxel_size, factors))
-    if activation is not None:
-        activation = getattr(tf.nn, activation)
-
-    if constant_upsample:
-
-        in_shape = tuple(fmaps_in.get_shape().as_list())
-        num_fmaps_in = in_shape[1]
-        num_fmaps_out = num_fmaps
-        out_shape = (
-            in_shape[0],
-            num_fmaps_out) + tuple(s*f for s, f in zip(in_shape[2:], factors))
-
-        if is_2d:
-            # (num_fmaps_out * num_fmaps_in)
-            kernel_variables = tf.get_variable(
-                name + '_kernel_variables',
-                (num_fmaps_out * num_fmaps_in,),
-                dtype=tf.float32)
-            # (1, 1, num_fmaps_out, num_fmaps_in)
-            kernel_variables = tf.reshape(
-                kernel_variables,
-                (1, 1) + (num_fmaps_out, num_fmaps_in))
-            # (f_y, f_x, num_fmaps_out, num_fmaps_in)
-            constant_upsample_filter = repeat(
-                kernel_variables,
-                tuple(factors) + (1, 1))
-
-            fmaps = tf.nn.conv2d_transpose(
-                fmaps_in,
-                filter=constant_upsample_filter,
-                output_shape=out_shape,
-                strides=(1, 1) + tuple(factors),
-                padding='VALID',
-                data_format='NCHW',
-                name=name)
-
-            if activation is not None:
-                fmaps = activation(fmaps)
-
-        else:
-            # (num_fmaps_out * num_fmaps_in)
-            kernel_variables = tf.get_variable(
-                name + '_kernel_variables',
-                (num_fmaps_out * num_fmaps_in,),
-                dtype=tf.float32)
-            # (1, 1, 1, num_fmaps_out, num_fmaps_in)
-            kernel_variables = tf.reshape(
-                kernel_variables,
-                (1, 1, 1) + (num_fmaps_out, num_fmaps_in))
-            # (f_z, f_y, f_x, num_fmaps_out, num_fmaps_in)
-            constant_upsample_filter = repeat(
-                kernel_variables,
-                tuple(factors) + (1, 1))
-
-            fmaps = tf.nn.conv3d_transpose(
-                fmaps_in,
-                filter=constant_upsample_filter,
-                output_shape=out_shape,
-                strides=(1, 1) + tuple(factors),
-                padding='VALID',
-                data_format='NCDHW',
-                name=name)
-
-            if activation is not None:
-                fmaps = activation(fmaps)
-
-    else:
-        if is_2d:
-            fmaps = tf.layers.conv2d_transpose(
-                fmaps_in,
-                filters=num_fmaps,
-                kernel_size=factors,
-                strides=factors,
-                padding='valid',
-                data_format='channels_first',
-                activation=activation,
-                name=name,
-                )
-
-        else:
-            fmaps = tf.layers.conv3d_transpose(
-                fmaps_in,
-                filters=num_fmaps,
-                kernel_size=factors,
-                strides=factors,
-                padding='valid',
-                data_format='channels_first',
-                activation=activation,
-                name=name)
-
-    return fmaps, voxel_size
-
-
-def crop(fmaps_in, shape):
-    '''Crop spatial and time dimensions to match shape.
-
-    Args:
-
-        fmaps_in:
-
-            The input tensor of shape ``(b, c, z, y, x)`` (for 3D) or ``(b, c,
-            t, z, y, x)`` (for 4D).
-
-        shape:
-
-            A list (not a tensor) with the requested shape ``[_, _, z, y, x]``
-            (for 3D) or ``[_, _, t, z, y, x]`` (for 4D).
-    '''
-
-    in_shape = fmaps_in.get_shape().as_list()
-
-    # Explicitly handle number of dimensions
-    in_is_4d = len(in_shape) == 6
-    in_is_2d = len(in_shape) == 4
-    out_is_4d = len(shape) == 6
-
-    if in_is_4d and not out_is_4d:
-        # set output shape for time to 1
-        shape = shape[0:2] + [1] + shape[2:]
-
-    if in_is_4d:
-        offset = [
-            0,  # batch
-            0,  # channel
-            (in_shape[2] - shape[2])//2,  # t
-            (in_shape[3] - shape[3])//2,  # z
-            (in_shape[4] - shape[4])//2,  # y
-            (in_shape[5] - shape[5])//2,  # x
-        ]
-        size = [
-            in_shape[0],
-            in_shape[1],
-            shape[2],
-            shape[3],
-            shape[4],
-            shape[5],
-        ]
-    elif in_is_2d:
-        offset = [
-            0,  # batch
-            0,  # channel
-            (in_shape[2] - shape[2])//2,  # y
-            (in_shape[3] - shape[3])//2,  # x
-        ]
-        size = [
-            in_shape[0],
-            in_shape[1],
-            shape[2],
-            shape[3],
-        ]
-
-    else:
-        offset = [
-            0,  # batch
-            0,  # channel
-            (in_shape[2] - shape[2])//2,  # z
-            (in_shape[3] - shape[3])//2,  # y
-            (in_shape[4] - shape[4])//2,  # x
-        ]
-        size = [
-            in_shape[0],
-            in_shape[1],
-            shape[2],
-            shape[3],
-            shape[4],
-        ]
-
-    fmaps = tf.slice(fmaps_in, offset, size)
-
-    if in_is_4d and not out_is_4d:
-        # remove time dimension
-        shape = shape[0:2] + shape[3:]
-        fmaps = tf.reshape(fmaps, shape)
-
-    return fmaps
-
-
-def crop_to_factor(fmaps_in, factor, kernel_sizes):
-    '''Crop feature maps to ensure translation equivariance with stride of
-    upsampling factor. This should be done right after upsampling, before
-    application of the convolutions with the given kernel sizes.
-
-    The crop could be done after the convolutions, but it is more efficient to
-    do that before (feature maps will be smaller).
-    '''
-
-    shape = fmaps_in.get_shape().as_list()
-    spatial_dims = len(shape) - 2
-    spatial_shape = shape[-spatial_dims:]
-
-    # the crop that will already be done due to the convolutions
-    convolution_crop = list(
-        sum(
-            (ks if isinstance(ks, int) else ks[d]) - 1
-            for ks in kernel_sizes
-        )
-        for d in range(spatial_dims)
-    )
-    print("crop_to_factor: factor =", factor)
-    print("crop_to_factor: kernel_sizes =", kernel_sizes)
-    print("crop_to_factor: convolution_crop =", convolution_crop)
-
-    # we need (spatial_shape - convolution_crop) to be a multiple of factor,
-    # i.e.:
-    #
-    # (s - c) = n*k
-    #
-    # we want to find the largest n for which s' = n*k + c <= s
-    #
-    # n = floor((s - c)/k)
-    #
-    # this gives us the target shape s'
-    #
-    # s' = n*k + c
-
-    ns = (
-        int(math.floor(float(s - c)/f))
-        for s, c, f in zip(spatial_shape, convolution_crop, factor)
-    )
-    target_spatial_shape = tuple(
-        n*f + c
-        for n, c, f in zip(ns, convolution_crop, factor)
-    )
-
-    if target_spatial_shape != spatial_shape:
-
-        assert all((
-                (t > c) for t, c in zip(
-                    target_spatial_shape,
-                    convolution_crop))
-            ), \
-            "Feature map with shape %s is too small to ensure translation " \
-            "equivariance with factor %s and following convolutions %s" % (
-                shape,
-                factor,
-                kernel_sizes)
-
-        target_shape = list(shape)
-        target_shape[-spatial_dims:] = target_spatial_shape
-
-        print("crop_to_factor: shape =", shape)
-        print("crop_to_factor: spatial_shape =", spatial_shape)
-        print("crop_to_factor: target_spatial_shape =", target_spatial_shape)
-        print("crop_to_factor: target_shape =", target_shape)
-        fmaps = crop(
-            fmaps_in,
-            target_shape)
-    else:
-        fmaps = fmaps_in
-
-    return fmaps
+from .layers import conv_pass, downsample, upsample
+from .utils import (crop_spatial_temporal, crop_to_factor,
+                    get_number_of_tf_variables)
+
+logger = logging.getLogger(__name__)
 
 
 def unet(
@@ -712,8 +226,9 @@ def unet(
                 str(g_out_upsampled.shape))
 
             # copy-crop
-            f_left_cropped = crop(f_left,
-                                  g_out_upsampled.get_shape().as_list())
+            f_left_cropped = crop_spatial_temporal(
+                f_left,
+                g_out_upsampled.get_shape().as_list())
 
             print(prefix + "f_left_cropped: " + str(f_left_cropped.shape))
